+++ packages/market-service/src/foxy/foxy.ts
@@ -1,4 +1,8 @@
+import { ethChainId } from '@shapeshiftoss/caip'
+import { ChainAdapter, ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
+import { foxyAddresses, FoxyApi } from '@shapeshiftoss/investor-foxy'
 import {
+  ChainTypes,
   HistoryData,
   HistoryTimeframe,
   MarketCapResult,
@@ -11,19 +15,36 @@ import dayjs from 'dayjs'
 import { MarketService } from '../api'
 import { CoinCapMarketCap } from '../coincap/coincap-types'
 import { RATE_LIMIT_THRESHOLDS_PER_MINUTE } from '../config'
+import { ProviderUrls } from '../market-service-manager'
 import { bn } from '../utils/bignumber'
 import { isValidDate } from '../utils/isValidDate'
 import { rateLimitedAxios } from '../utils/rateLimiters'
 
 export const FOXY_ASSET_ID = 'eip155:1/erc20:0xDc49108ce5C57bc3408c3A5E95F3d864eC386Ed3'
 const FOX_COINCAP_ID = 'fox-token'
+const FOXY_ASSET_PRECISION = '18'
 
 const axios = rateLimitedAxios(RATE_LIMIT_THRESHOLDS_PER_MINUTE.COINCAP)
 
 export class FoxyMarketService implements MarketService {
+  jsonRpcProviderUrl: string
+  adapterManager: ChainAdapterManager
   baseUrl = 'https://api.coincap.io/v2'
 
-  findAll = async () => {
+  constructor(providerUrls: ProviderUrls) {
+    this.jsonRpcProviderUrl = providerUrls.jsonRpcProviderUrl
+
+    const unchainedUrls = {
+      [ChainTypes.Ethereum]: {
+        // from web env, both are always defined despite what the typings suggest
+        httpUrl: providerUrls.unchainedEthereumHttpUrl,
+        wsUrl: providerUrls.unchainedEthereumWsUrl
+      }
+    }
+    this.adapterManager = new ChainAdapterManager(unchainedUrls)
+  }
+
+  async findAll() {
     try {
       const assetId = FOXY_ASSET_ID
       const marketData = await this.findByAssetId({ assetId })
@@ -35,7 +56,7 @@ export class FoxyMarketService implements MarketService {
     }
   }
 
-  findByAssetId = async ({ assetId }: MarketDataArgs): Promise<MarketData | null> => {
+  async findByAssetId({ assetId }: MarketDataArgs): Promise<MarketData | null> {
     try {
       if (assetId.toLowerCase() !== FOXY_ASSET_ID.toLowerCase()) {
         console.warn('FoxyMarketService(findByAssetId): Failed to find by AssetId')
@@ -45,11 +66,23 @@ export class FoxyMarketService implements MarketService {
       const { data } = await axios.get(`${this.baseUrl}/assets/${FOX_COINCAP_ID}`)
       const marketData = data.data as CoinCapMarketCap
 
+      // Make maxSupply as an additional field, effectively EIP-20's totalSupply
+      const api = new FoxyApi({
+        adapter: this.adapterManager.byChainId(ethChainId) as ChainAdapter<ChainTypes.Ethereum>,
+        providerUrl: this.jsonRpcProviderUrl,
+        foxyAddresses
+      })
+      const tokenContractAddress = foxyAddresses[0].foxy
+      const foxyTotalSupply = await api.totalSupply({ tokenContractAddress })
+      const supply = await api.tvl({ tokenContractAddress })
+
       return {
         price: marketData.priceUsd,
         marketCap: '0', // TODO: add marketCap once able to get foxy marketCap data
         changePercent24Hr: parseFloat(marketData.changePercent24Hr),
-        volume: '0' // TODO: add volume once able to get foxy volume data
+        volume: '0', // TODO: add volume once able to get foxy volume data
+        supply: supply?.div(`1e+${FOXY_ASSET_PRECISION}`).toString(),
+        maxSupply: foxyTotalSupply?.div(`1e+${FOXY_ASSET_PRECISION}`).toString()
       }
     } catch (e) {
       console.warn(e)
@@ -57,10 +90,10 @@ export class FoxyMarketService implements MarketService {
     }
   }
 
-  findPriceHistoryByAssetId = async ({
+  async findPriceHistoryByAssetId({
     assetId,
     timeframe
-  }: PriceHistoryArgs): Promise<HistoryData[]> => {
+  }: PriceHistoryArgs): Promise<HistoryData[]> {
     if (assetId.toLowerCase() !== FOXY_ASSET_ID.toLowerCase()) {
       console.warn(
         'FoxyMarketService(findPriceHistoryByAssetId): Failed to find price history by AssetId'