+++ packages/swapper/src/manager/SwapperManager.test.ts
@@ -20,7 +20,8 @@ describe('SwapperManager', () => {
   }
 
   const thorchainSwapperDeps: ThorchainSwapperDeps = {
-    midgardUrl: 'localhost:3000'
+    midgardUrl: 'localhost:3000',
+    adapterManager: <ChainAdapterManager>{}
   }
 
   describe('constructor', () => {